Create triangle.html
Frontend Triangle Challenge:
Write a program that will determine the type of a triangle. It should take the lengths of the triangle's three sides as input, and return whether the triangle is equilateral, isosceles or scalene. A user should be able to provide inputs to determine the type of a triangle, either via a web UI or command line arguments. We are looking for solutions that showcase problem solving skills and structural considerations that can be applied to larger and potentially more complex problem domains. Pay special attention to tests, readability of code and error cases.
The resulting code will be used as the basis for broader discussions during the interview. Therefore you should be prepared to discuss your choices.